requirements.txt
evaluate==0.3.0
hydra-core==1.3.1
jsonlines==3.1.0
matplotlib==3.6.1
numpy==1.23.4
omegaconf==2.2.3
pandas==1.5.1
scikit_learn==1.2.0
scipy==1.10.0
texttable==1.6.4
torch==1.12.1+cu113
tqdm==4.64.1
transformers==4.23.1
